Nozomi.Ticker: Startup: NozomiAuthContext: EnableSensitiveDataLogging
This resolves and deepens this error message:
InvalidOperationException: The instance of entity type 'User' cannot be tracked because another instance with the same key value for {'Id'} is already being tracked. When attaching existing entities, ensure that only one entity instance with a given key value is attached. Consider using 'DbContextOptionsBuilder.EnableSensitiveDataLogging' to see the conflicting key values.

Signed-off-by: nixxholas <nicholas@slyck.net>